fix(preview): gate the "Full scaffold" label on real source, not confidence alone
readScaffoldPreview labelled a strong blueprint selection "Full scaffold" even
when the scaffold emitted no source code (only a manifest plus empty dirs). PR
#82 already made the post-scaffold review verdict honest with a source-file
signal, but the preview label was never updated to match, so it kept
overclaiming.

readScaffoldPreview now also gates FULL_SCAFFOLD on real runtime structure,
reusing the exact signal #82 added (a generated source file by extension). To
avoid a circular import (post-scaffold-review.ts already imports from
planforge-output.ts), the shared detection (PLANFORGE_TOP_LEVEL_PATHS,
SOURCE_FILE_EXTENSIONS, collectRuntimePaths, hasRuntimeSignals) moves to a new
lib/runtime-signals.ts that both modules import, so the label and the verdict
have a single source of truth.

Tests: a strong-confidence scaffold with no source is now planning-baseline (the
regression), a strong-confidence scaffold with real source is full, and weak
confidence stays planning-baseline even with source present (confidence gate
wins).

Refs: agent-tasks 42b0fbe6
